+++ examples/validation_framework.py
@@ -80,8 +80,8 @@ import pytesmo.validation_framework.metric_calculators as metrics_calculators
 
 from datetime import datetime
 
-from ascat.timeseries import AscatSsmCdr
-from pytesmo.io.ismn.interface import ISMN_Interface
+from ascat.read_native.cdr import AscatSsmCdr
+from ismn.interface import ISMN_Interface
 from pytesmo.validation_framework.validation import Validation
 from pytesmo.validation_framework.results_manager import netcdf_results_manager
 
@@ -256,7 +256,7 @@ results_fname = os.path.join(save_path, 'ASCAT.sm_with_ISMN.soil moisture.nc')
 
 with netCDF4.Dataset(results_fname) as ds:
     for var in ds.variables:
-        print var, ds.variables[var][:]
+        print(var, ds.variables[var][:])
 
 
 # ## Parallel processing
@@ -313,7 +313,7 @@ def start_processing(job):
 from pytesmo.validation_framework.adapters import MaskingAdapter
 
 ds_mask = MaskingAdapter(ismn_reader, '<', 0.2, 'soil moisture')
-print ds_mask.read_ts(ids[0]).head()
+print(ds_mask.read_ts(ids[0]).head())
 
 
 # ### Self-masking adapter
@@ -325,5 +325,5 @@ print ds_mask.read_ts(ids[0]).head()
 from pytesmo.validation_framework.adapters import SelfMaskingAdapter
 
 ds_mask = SelfMaskingAdapter(ismn_reader, '<', 0.2, 'soil moisture')
-print ds_mask.read_ts(ids[0]).head()
+print(ds_mask.read_ts(ids[0]).head())