+++ pytesmo/time_series/anomaly.py
@@ -93,7 +93,11 @@ def calc_climatology(Ser,
                      timespan=None,
                      fill=np.nan,
                      wraparound=False,
-                     respect_leap_years=False):
+                     respect_leap_years=False,
+                     interpolate_leapday=False,
+                     fillna=True,
+                     min_obs_orig=1,
+                     min_obs_clim=1):
     '''
     Calculates the climatology of a data set.
 
@@ -130,6 +134,18 @@ def calc_climatology(Ser,
         the climatology
         Default: False
 
+    fillna: boolean, optional
+        If set, then the moving average used for the calculation of the
+        climatology will be filled at the nan-values
+
+    min_obs_orig: int
+        Minimum observations required to give a valid output in the first
+        moving average applied on the input series
+
+    min_obs_clim: int
+        Minimum observations required to give a valid output in the second
+        moving average applied on the calculated climatology
+
     Returns
     -------
     climatology : pandas.Series
@@ -140,7 +156,7 @@ def calc_climatology(Ser,
     if timespan is not None:
         Ser = Ser.truncate(before=timespan[0], after=timespan[1])
 
-    Ser = moving_average(Ser, window_size=moving_avg_orig)
+    Ser = moving_average(Ser, window_size=moving_avg_orig, fillna=fillna, min_obs=min_obs_orig)
 
     Ser = pd.DataFrame(Ser)
 
@@ -151,10 +167,14 @@ def calc_climatology(Ser,
     else:
         year, month, day = julian2date(Ser.index.values)[0:3]
 
+
+
+
     if respect_leap_years:
         doys = doy(month, day, year)
     else:
         doys = doy(month, day)
+
 
     Ser['doy'] = doys
 
@@ -165,6 +185,11 @@ def calc_climatology(Ser,
 
     clim_ser = pd.Series(clim.values.flatten(),
                          index=clim.index.values)
+
+    if interpolate_leapday and not respect_leap_years:
+        clim_ser[60] = np.mean((clim_ser[59], clim_ser[61]))
+    elif interpolate_leapday and respect_leap_years:
+        clim_ser[366] = np.mean((clim_ser[365], clim_ser[1]))
 
     if wraparound:
         index_old = clim_ser.index.copy()
@@ -178,11 +203,11 @@ def calc_climatology(Ser,
                               clim_ser,
                               right_mirror])
 
-        clim_ser = moving_average(clim_ser, window_size=moving_avg_clim)
+        clim_ser = moving_average(clim_ser, window_size=moving_avg_clim, fillna=fillna, min_obs=min_obs_clim)
         clim_ser = clim_ser.iloc[moving_avg_clim:-moving_avg_clim]
         clim_ser.index = index_old
     else:
-        clim_ser = moving_average(clim_ser, window_size=moving_avg_clim)
+        clim_ser = moving_average(clim_ser, window_size=moving_avg_clim, fillna=fillna, min_obs=min_obs_clim)
 
     clim_ser = clim_ser.reindex(np.arange(366) + 1)
     clim_ser = clim_ser.fillna(fill)